chore: move email page in dedicated iframe
